'just build-and-run' command
This makes it easy to build a sample app and run it in serval with a single command.

If your sample project emits an artifact with the same name as the project, you can just do this:

````sh
just build-and-run sdk-test
````

This will do `just build sdk-test` and then run `./build/sdk-test.wasm` in Serval.

If your project emits multiple WASM files or creates an artifact with a different name than the
project, you can specify the .wasm binary you want to run as a second parameter:

````sh
just build-and-run assemblyscript build/as-hello.wasm
````